refactor pixel tree to package
