+++ src/hipscat/catalog/__init__.py
@@ -3,5 +3,3 @@
 from .catalog import Catalog
 from .catalog_parameters import CatalogParameters
 from .partition_info import PartitionInfo
-from .pixel_node import PixelNode
-from .pixel_node_type import PixelNodeType
