+++ src/hipscat/pixel_tree/pixel_node_type.py
